- Changed class Deck so the constructor takes a byte for argument instead of an integer. The purpose of the change is to make it more obvious that the argument is for the amount of sets and not the amount of cards. - Game: 	- All fields in the class are now private. 	- Method newDeck is now completed. 	- initateGame has been renamed to restartGame, in order to not 	confuse it with the intent of the constructor. - GameTest: 	- Removed several testcases that were either some form of duplicates, or testing things that were not in the scope of this project. 	- Changed newDeckTest to compare for size of the deck rather 	than just compare for null.

+++ src/cards/Deck.java
@@ -10,14 +10,14 @@ public class Deck
 	
 	public Deck()
 	{
-		this(1);
+		this((byte) 1);
 	}
 	
-	public Deck(int numberOfSets)
+	public Deck(byte numberOfSets)
 	{
 		if(numberOfSets < 1)
 			throw new IllegalArgumentException("The deck must consist of at least one complete set.");
-		for(int i = 0; i < numberOfSets; i++)
+		for(byte i = 0; i < numberOfSets; i++)
 			generateCards();
 	}
